client-deploy/ailinux-android-app/requirements.txt
# AILinux Android Client Dependencies
kivy==2.3.0
kivymd==1.2.0
httpx>=0.25.0
certifi
pillow
